Change scope of fixture from module to session
